--- app/src/main/java/com/kionavani/todotask/ui/viewmodels/ToDoViewModelFactory.kt
@@ -1,15 +1,0 @@
-package com.kionavani.todotask.ui.viewmodels
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.kionavani.todotask.data.TodoItemsRepository
-
-class ToDoViewModelFactory(private val repository: TodoItemsRepository) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(ToDoViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return ToDoViewModel(repository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
